Switched Repository primary key type from Integer to Long

+++ src/main/java/com/mellora/drugconsortium/dao/CompanyRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.mellora.drugconsortium.model.Company;
 
-public interface CompanyRepo extends JpaRepository<Company, Integer> {
+public interface CompanyRepo extends JpaRepository<Company, Long> {
 
 }
